+++ backend/config/api.py
@@ -8,6 +8,7 @@ comando de contrato funcione tal cual:
 from ninja import NinjaAPI
 
 from apps.core.api import router as core_router
+from apps.geo.api import router as geo_router
 
 api = NinjaAPI(
     title="ConstruScan API",
@@ -16,3 +17,4 @@ api = NinjaAPI(
 )
 
 api.add_router("", core_router)
+api.add_router("", geo_router)